@@ -1,17 +1,19 @@
-require('dotenv').config();
-const { Sequelize } = require('sequelize');
+import { Sequelize } from "sequelize";
+import "dotenv/config";
 
-
-const sequelize = new Sequelize(
-  process.env.DB_NAME,
-  process.env.DB_USER,
-  process.env.DB_PASS,
+const db = new Sequelize(
+  // process.env.DB_NAME,
+  // process.env.DB_USER,
+  // process.env.DB_PASS,
+  "ecowaste",
+  "root",
+  "",
   {
     host: process.env.DB_HOST || 'localhost',
-    port: 3306,
+    // port: 3306,
     dialect: 'mysql',
     logging: false,
   }
 );
 
-module.exports = sequelize;
+export default db;
