@@ -1,45 +1,44 @@
-const { DataTypes } = require('sequelize');
-const sequelize = require('../config/db');
+import db from "../config/db.js";
+import { Sequelize } from "sequelize";
 
-const Location = sequelize.define('Location', {
-  id: {
-    type: DataTypes.INTEGER,
-    primaryKey: true,
-    autoIncrement: true,
-  },
-  name: {
-    type: DataTypes.STRING(100),
-    allowNull: false,
-  },
-  province: {
-    type: DataTypes.STRING(50),
-    allowNull: false,
-  },
-  city: {
-    type: DataTypes.STRING(50),
-    allowNull: false,
-  },
-  address: {
-    type: DataTypes.STRING(255),
-    allowNull: true,
-  },
-  latitude: {
-    type: DataTypes.DOUBLE,
-    allowNull: false,
-  },
-  longitude: {
-    type: DataTypes.DOUBLE,
-    allowNull: false,
-  },
-  waste_type: {
-    type: DataTypes.ENUM('Plastik', 'Kertas', 'Logam', 'Organik', 'Elektronik', 'B3'),
-    allowNull: false,
-  },
-}, {
-  tableName: 'locations',
-  timestamps: true, // karena kolom created_at manual
-  underscored: true,
+const Location = db.define("Location", {
+    id: {
+        type: Sequelize.INTEGER,
+        primaryKey: true,
+        autoIncrement: true,
+    },
+    name: {
+        type: Sequelize.STRING(100),
+        allowNull: false,
+    },
+    province: {
+        type: Sequelize.STRING(50),
+        allowNull: false,
+    },
+    city: {
+        type: Sequelize.STRING(50),
+        allowNull: false,
+    },
+    address: {
+        type: Sequelize.STRING(255),
+        allowNull: true,
+    },
+    latitude: {
+        type: Sequelize.DOUBLE,
+        allowNull: false,
+    },
+    longitude: {
+        type: Sequelize.DOUBLE,
+        allowNull: false,
+    },
+    waste_type: {
+        type: Sequelize.ENUM('Plastik', 'Kertas', 'Logam', 'Organik', 'Elektronik', 'B3'),
+        allowNull: false,
+    },
+    waste_id: {
+        type: Sequelize.INTEGER,
+        allowNull: true,
+    },
 });
 
-module.exports = Location;
-
+export { Location };
